Posterior simulation in the panel probit model with general sigma_2
This is an extension of 14.4 in which the sigma_alpha_2 is equal 1 but sigma_2 is general parameter, whose posterior is sampled as well.
In more general setup both sigma_alpha_2 and sigma_2 shall be general parameters to be estimated. 

Case analyzed in this post assumes that eps ~ N(0, sigma_2) and we estimate sigma_2, being identical for every disturbance in the model, irrespective of the individual. The more general model can allow for so called groupwise heteroscedasticity, in which every individual's latent process is characterized by different value of sigma2.