Posterior sampling in the panel ordered probit model
Code is based on the mix of implementations in ex_14_4_ sigma2_alpha and ex_14_5_cutpoints_step.

This way we obtain the complexity of panel and ordered probit at the same time.

The implementation assumes fixed sigma2 parameter.
This is a natural extension.

Also, blockwise heteroscedasticity structured, with different simga2 for each individual could be implemented.